// vehicle-service-management/src/app/core/api/generated/types.gen.ts
export type CustomerSummaryDto = {
  id: number | string;
  fullName: string;
  isArchived: boolean;
};
// ...
export type ServiceRecordDto = {
  id: number | string;
  serviceTitle: string;
  serviceType: string;
  description: string;
  laborCost: number | string;
  partsCost: number | string;
  totalCost: number | string;
  status: string;
  serviceDate: string;
  mileageAtService: number | string;
  vehicleId: number | string;
  vehicle?: null | ServiceRecordVehicleDto;
  createdAt: string;
  updatedAt: string;
  isArchived: boolean;
  rowVersion: string;
};

export type ServiceRecordVehicleDto = {
  id: number | string;
  plateNumber: string;
  brand: string;
  model: string;
  year: number | string;
  color: string;
  vin?: null | string;
  currentMileage: number | string;
  isArchived: boolean;
  customer?: null | CustomerSummaryDto;
};
// ...
export type VehicleDto = {
  id: number | string;
  plateNumber: string;
  brand: string;
  model: string;
  year: number | string;
  color: string;
  vin?: null | string;
  currentMileage: number | string;
  customerId?: null | number | string;
  customer?: null | CustomerSummaryDto;
  createdAt: string;
  updatedAt: string;
  isArchived: boolean;
  rowVersion: string;
};